% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{analyze_basket.basket_design}
\alias{analyze_basket.basket_design}
\title{Basket design analyze method}
\usage{
\method{analyze_basket}{basket_design}(data, design, ...)
}
\arguments{
\item{data}{A basket_data object}

\item{design}{A basket_design object}

\item{...}{Additional arguments passed to method-specific functions}
}
\description{
Basket design analyze method
}
