% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{analyze_basket.default}
\alias{analyze_basket.default}
\title{Default analyze_basket method}
\usage{
\method{analyze_basket}{default}(data, design, ...)
}
\arguments{
\item{data}{A basket_data object}

\item{design}{A design object}

\item{...}{Additional arguments}
}
\description{
Default analyze_basket method
}
