% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{analyze_basket}
\alias{analyze_basket}
\title{Analyze simulated basket trial}
\usage{
analyze_basket(data, design, ...)
}
\arguments{
\item{data}{A basket_data object or list of basket_data objects}

\item{design}{A basket_design object specifying the analysis method}

\item{...}{Additional arguments passed to the specific analysis function}
}
\value{
An analysis results object
}
\description{
Analyzes basket trial data using the specified design method
}
