% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket_trial.R
\name{basket_data}
\alias{basket_data}
\title{Create a basket trial data object}
\usage{
basket_data(basket_names, n_patients, n_responses)
}
\arguments{
\item{basket_names}{Character vector. Names of the baskets}

\item{n_patients}{Integer vector. Number of patients per basket}

\item{n_responses}{Integer vector. Number of responses per basket}
}
\value{
An object of class 'basket_data' containing the trial data
}
\description{
Creates a structured data object representing observed data from a basket trial
}
\examples{
# Imatinib trial data from Chugh et al. 2009
data <- basket_data(
  basket_names = c("Angiosarcoma", "Ewing", "Fibrosarcoma"),
  n_patients = c(15, 13, 12),
  n_responses = c(2, 0, 1)
)
}
