% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_sample_size}
\alias{calculate_sample_size}
\title{Calculate sample size for basket trial}
\usage{
calculate_sample_size(
  n_baskets,
  null_rate,
  alt_rate,
  power = 0.8,
  alpha = 0.05,
  method = "bma"
)
}
\arguments{
\item{n_baskets}{Integer. Number of baskets}

\item{null_rate}{Numeric. Null response rate}

\item{alt_rate}{Numeric. Alternative response rate}

\item{power}{Numeric. Desired power (0-1)}

\item{alpha}{Numeric. Significance level}

\item{method}{Character. Design method to use}
}
\value{
Recommended sample size per basket
}
\description{
Provides sample size recommendations based on desired operating characteristics
}
\examples{
# Calculate sample size for BMA design
n <- calculate_sample_size(
  n_baskets = 4,
  null_rate = 0.20,
  alt_rate = 0.35,
  power = 0.80,
  alpha = 0.05,
  method = "bma"
)
}
