% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_designs}
\alias{compare_designs}
\title{Compare operating characteristics across designs}
\usage{
compare_designs(..., metrics = c("power", "type_i_error", "fwer"))
}
\arguments{
\item{...}{simulation_results objects to compare}

\item{metrics}{Character vector of metrics to compare}
}
\value{
A comparison table
}
\description{
Compares operating characteristics from multiple simulation results
}
