% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{compute_operating_characteristics}
\alias{compute_operating_characteristics}
\title{Compute operating characteristics from simulation results}
\usage{
compute_operating_characteristics(results, design, alpha = 0.05)
}
\arguments{
\item{results}{List of analysis results from simulations}

\item{design}{Original basket_design object}

\item{alpha}{Significance level}
}
\value{
A list containing operating characteristics
}
\description{
Calculates type I error, power, FWER, etc. from simulation results
}
