% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_bhm.R
\name{extract_rejections.bhm_result}
\alias{extract_rejections.bhm_result}
\title{Extract rejections from BHM results}
\usage{
\method{extract_rejections}{bhm_result}(result, ...)
}
\arguments{
\item{result}{A bhm_result object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Logical vector of rejection decisions
}
\description{
Extract rejections from BHM results
}
