% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{extract_rejections.default}
\alias{extract_rejections.default}
\title{Default extract_rejections method}
\usage{
\method{extract_rejections}{default}(result, ...)
}
\arguments{
\item{result}{Analysis result object}

\item{...}{Additional arguments (not currently used)}
}
\description{
Default extract_rejections method
}
