% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_decision_criteria_text}
\alias{generate_decision_criteria_text}
\title{Generate decision criteria text}
\usage{
generate_decision_criteria_text(design)
}
\arguments{
\item{design}{A basket_design object}
}
\description{
Generate decision criteria text
}
\keyword{internal}
