% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_design_summary}
\alias{generate_design_summary}
\title{Generate design summary section}
\usage{
generate_design_summary(design, format)
}
\description{
Generate design summary section
}
\keyword{internal}
