% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_hypothesis_text}
\alias{generate_hypothesis_text}
\title{Generate hypothesis testing text}
\usage{
generate_hypothesis_text(design)
}
\arguments{
\item{design}{A basket_design object}
}
\description{
Generate hypothesis testing text
}
\keyword{internal}
