% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_interim_summary}
\alias{generate_interim_summary}
\title{Generate interim analysis summary}
\usage{
generate_interim_summary(result, format)
}
\description{
Generate interim analysis summary
}
\keyword{internal}
