% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_references}
\alias{generate_references}
\title{Generate references}
\usage{
generate_references(design)
}
\arguments{
\item{design}{A basket_design object}
}
\description{
Generate references
}
\keyword{internal}
