% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_rmarkdown_protocol}
\alias{generate_rmarkdown_protocol}
\title{Generate R Markdown protocol with reproducible code}
\usage{
generate_rmarkdown_protocol(design, ...)
}
\description{
Generate R Markdown protocol with reproducible code
}
\keyword{internal}
