% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_sap_section}
\alias{generate_sap_section}
\title{Generate Statistical Analysis Plan section}
\usage{
generate_sap_section(design)
}
\description{
Generate Statistical Analysis Plan section
}
\keyword{internal}
