% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_statistical_approach_text}
\alias{generate_statistical_approach_text}
\title{Generate statistical approach text}
\usage{
generate_statistical_approach_text(design, include_details)
}
\arguments{
\item{design}{A basket_design object}

\item{include_details}{Logical}
}
\description{
Generate statistical approach text
}
\keyword{internal}
