% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_technical_details}
\alias{generate_technical_details}
\title{Generate technical details}
\usage{
generate_technical_details(design)
}
\arguments{
\item{design}{A basket_design object}
}
\description{
Generate technical details
}
\keyword{internal}
