% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_language.R
\name{generate_word_protocol}
\alias{generate_word_protocol}
\title{Generate Word document protocol}
\usage{
generate_word_protocol(design, file, ...)
}
\description{
Generate Word document protocol
}
\keyword{internal}
