% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{get_decision_threshold}
\alias{get_decision_threshold}
\title{Get decision threshold for a basket}
\usage{
get_decision_threshold(interim_result, design, basket_idx)
}
\description{
Get decision threshold for a basket
}
\keyword{internal}
