% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket_trial.R
\name{print.basket_data}
\alias{print.basket_data}
\title{Print method for basket_data}
\usage{
\method{print}{basket_data}(x, ...)
}
\arguments{
\item{x}{A basket_data object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for basket_data
}
