% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket_trial.R
\name{print.basket_design}
\alias{print.basket_design}
\title{Print method for basket_design}
\usage{
\method{print}{basket_design}(x, ...)
}
\arguments{
\item{x}{A basket_design object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for basket_design
}
