% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_bma.R
\name{print.bma_result}
\alias{print.bma_result}
\title{Print method for BMA results}
\usage{
\method{print}{bma_result}(x, ...)
}
\arguments{
\item{x}{A bma_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for BMA results
}
