% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{print.conditional_power_result}
\alias{print.conditional_power_result}
\title{Print conditional power results}
\usage{
\method{print}{conditional_power_result}(x, ...)
}
\arguments{
\item{x}{A conditional_power_result object}

\item{...}{Additional arguments (not used)}
}
\description{
Print conditional power results
}
