% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_cunanan.R
\name{print.cunanan_result}
\alias{print.cunanan_result}
\title{Print method for Cunanan results}
\usage{
\method{print}{cunanan_result}(x, ...)
}
\arguments{
\item{x}{A \code{cunanan_result} object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for Cunanan results
}
