% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.design_comparison}
\alias{print.design_comparison}
\title{Print method for design_comparison}
\usage{
\method{print}{design_comparison}(x, ...)
}
\arguments{
\item{x}{A design_comparison object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for design_comparison
}
