% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mem.R
\name{print.mem_result}
\alias{print.mem_result}
\title{Print method for MEM results}
\usage{
\method{print}{mem_result}(x, ...)
}
\arguments{
\item{x}{A mem_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for MEM results
}
