% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_simon.R
\name{print.simon_result}
\alias{print.simon_result}
\title{Print method for Simon results}
\usage{
\method{print}{simon_result}(x, ...)
}
\arguments{
\item{x}{A \code{simon_result} object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for Simon results
}
