% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.simulation_results}
\alias{print.simulation_results}
\title{Print method for simulation_results}
\usage{
\method{print}{simulation_results}(x, ...)
}
\arguments{
\item{x}{A simulation_results object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for simulation_results
}
